requirements.txt
#opencv-python==4.5.5.64 
tensorflow==2.8.0
streamlit==1.9.2
numpy
Pillow
#opencv-python-headless==4.5.5.64 